Control_gui_with controller is the old control_gui, but all commands are sent through adr_controller. I tested the whole system, with the gui calling commands through the adr_controller and fetching data and the server using sockets with locks to safely grab data from the sim900.
The data is correctly transferred to the gui. Currently, regenerate turns the PID enabling on or off. This state can be set from the gui, and the system can manage to constantly fetch data while also sending commands to the sim900. The next step is actually programming useful steps in as well as the necessary safety checks. AdrController is really just a skeleton right now.