Wrote new programs fake_sim900 and test_server for controlled testing without using the real sim900.
Decided on a format for the input dictionary of commands (so the program can be used with different sim900's). Locked in on an output format (I will need to change the output of the socket server). Note that the get_data method is no longer used - the query requests are given from the command_dictionary.

Some general cleaning as well.