Continue testing on working ramp methods. Designed controller_ramp_test_script to test methods thoroughly.
Fixed controller toggling. Now ends/starts the loop thread, rather than setting controller=None (and not guarateeing the end of the loop_thread). Fixed GUI as well from checkbox to a button and a text box, since the checkbox could be manipulated so that it didn't show the state of the system.

Did some work on new regulate method.