Implemented regulate for the PID ramp control. Dwell now also stays in PID control rather than switching to manual control.
Improved logging for the sim900_server. State changes are now logged, as well as all sends and queries. Timing and flags have been moved down to DEBUG.

Control_gui now keeps track of an unlimited number of points for it's active (view on plots) lists, and 1000 for the inactive lists. Clear history resets these lists to [].

Added x-axis sharing for the plots.